junyoung 0203 (13251) 조합, 확률
